+++ CHANGELOG.md
@@ -8,12 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Sorting data sources
 - Table event OnDataCompare: 
   - Ability use the token in the value
   - Ability to use regular expressions in the column name
 
 ### Changed
 - removed libraries yFiles and license from project
+- Data source design
 
 ### Fixed
 - Risk-review: Fixed visible single bar