+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Sorting data sources
 - Mouse middle click by group tile and dashboard tile
 - Table event OnDataCompare: 
   - Ability use the token in the value
@@ -15,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - removed libraries yFiles and license from project
+- Data source design
 
 ### Fixed
 - Risk-review: Fixed visible single bar