+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+### Fixed
+
+## [1.147.0]
+
+### Added
 - Sorting data sources
 - Mouse middle click by group tile and dashboard tile
 - Table event OnDataCompare: